wakelock with timeout under-locked exception risk
When a wakelock is acquired with a timeout
and is released explicitely before the timeout ends,
it will issue a java runtime under-locked exception.
It happens because the releaser doesn't check
if the wakelock is still held before releasing it.

The fix consists in having the releaser check
if the wakelock is still held before releasing
it in an atomic way to prevent explicit release
concurrency between isHeld() and release() calls.
To be fully thread safe the explicit release needs
to use safeRelease() method as well.

Change-Id: I6de27a5aa8931187877b13e30c8a9a069b7bd622
Author: Fabien Peix <fabienx.peix@intel.com>
Signed-off-by: Ovidiu Beldie <ovidiux.beldie@intel.com>
Singed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 6705 17106