To fix the green title edge problem (JNI)
The problem was caused by alpha blending between some useless
green lines in the title image and the input image. In the
patch, we correct this wrong title boundary and only perform
blending between input image and title area.
There is also a very great side effect that: the computation
complexity is significantly reduced. It saves the exporting
time up to 60% for 1080P video with title effect.

Note: This patch is related to a change in platforms/frameworks/av

Change-Id: Id9bc4984ff52b1e0c5134005cd156d6686bcb1e2
Author: Wei Feng <wei.feng@intel.com>
Signed-off-by: Wei Feng <wei.feng@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 39058